pyproject.toml
[project]
name = "rag-web-search-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.9.0",
    "firecrawl-py>=4.3.6",
    "flake8>=7.3.0",
    "ipykernel>=6.30.1",
    "isort>=6.1.0",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-pinecone>=0.2.12",
    "langchain-tavily>=0.2.12",
    "langchainhub>=0.1.21",
    "langsmith>=0.4.33",
    "openai>=2.2.0",
    "pylink>=0.3.3",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "streamlit>=1.50.0",
    "streamlit-chat>=0.1.1",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.15",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
